E_learn/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
bleach==6.2.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-genai==1.13.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Markdown==3.8
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-decouple==3.8
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
webencodings==0.5.1
websockets==15.0.1
